+++ backend/emm-sale/src/documentTest/java/com/emmsale/ActivityApiTest.java
@@ -13,6 +13,7 @@ import com.emmsale.activity.api.ActivityApi;
 import com.emmsale.activity.application.dto.ActivityAddRequest;
 import com.emmsale.activity.application.dto.ActivityResponse;
 import com.emmsale.activity.domain.ActivityType;
+import com.emmsale.admin.activity.api.AdminActivityApi;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -27,7 +28,7 @@ import org.springframework.restdocs.payload.ResponseFieldsSnippet;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 
-@WebMvcTest(ActivityApi.class)
+@WebMvcTest({ActivityApi.class, AdminActivityApi.class}) // TODO: 2023/11/18 Admin API Test 분리하기
 class ActivityApiTest extends MockMvcTestHelper {
 
   @Test
@@ -35,26 +36,26 @@ class ActivityApiTest extends MockMvcTestHelper {
   void findAll() throws Exception {
     // given
     final ResponseFieldsSnippet responseFields = PayloadDocumentation.responseFields(
-      fieldWithPath("[].id").type(JsonFieldType.NUMBER).description("activity id"),
-      fieldWithPath("[].activityType").type(JsonFieldType.STRING).description("activity 분류"),
-      fieldWithPath("[].name").type(JsonFieldType.STRING).description("activity 이름")
+        fieldWithPath("[].id").type(JsonFieldType.NUMBER).description("activity id"),
+        fieldWithPath("[].activityType").type(JsonFieldType.STRING).description("activity 분류"),
+        fieldWithPath("[].name").type(JsonFieldType.STRING).description("activity 이름")
     );
 
     final List<ActivityResponse> expected = List.of(
-      new ActivityResponse(1L, "동아리", "YAPP"),
-      new ActivityResponse(2L, "동아리", "DND"),
-      new ActivityResponse(3L, "동아리", "nexters"),
-      new ActivityResponse(4L, "컨퍼런스", "인프콘"),
-      new ActivityResponse(5L, "교육", "우아한테크코스"),
-      new ActivityResponse(6L, "직무", "Backend")
+        new ActivityResponse(1L, "동아리", "YAPP"),
+        new ActivityResponse(2L, "동아리", "DND"),
+        new ActivityResponse(3L, "동아리", "nexters"),
+        new ActivityResponse(4L, "컨퍼런스", "인프콘"),
+        new ActivityResponse(5L, "교육", "우아한테크코스"),
+        new ActivityResponse(6L, "직무", "Backend")
     );
 
     Mockito.when(activityQueryService.findAll()).thenReturn(expected);
 
     // when & then
     mockMvc.perform(MockMvcRequestBuilders.get("/activities"))
-      .andExpect(MockMvcResultMatchers.status().isOk())
-      .andDo(MockMvcRestDocumentation.document("find-all-activities", responseFields));
+        .andExpect(MockMvcResultMatchers.status().isOk())
+        .andDo(MockMvcRestDocumentation.document("find-all-activities", responseFields));
   }
 
   @Test
@@ -62,29 +63,29 @@ class ActivityApiTest extends MockMvcTestHelper {
   void addTag() throws Exception {
     //given
     final RequestFieldsSnippet requestFields = requestFields(
-      fieldWithPath("activityType").type(JsonFieldType.STRING).description("활동 유형"),
-      fieldWithPath("name").type(JsonFieldType.STRING).description("활동 이름")
+        fieldWithPath("activityType").type(JsonFieldType.STRING).description("활동 유형"),
+        fieldWithPath("name").type(JsonFieldType.STRING).description("활동 이름")
     );
 
     final ActivityAddRequest request = new ActivityAddRequest(ActivityType.CLUB, "DND");
     final ActivityResponse response = new ActivityResponse(3L,
-      ActivityType.CLUB.getValue(),
-      "DND"
+        ActivityType.CLUB.getValue(),
+        "DND"
     );
 
     when(activityCommandService.addActivity(any(ActivityAddRequest.class))).thenReturn(response);
 
     final ResponseFieldsSnippet responseFields = responseFields(
-      fieldWithPath("id").type(JsonFieldType.NUMBER).description("활동 식별자"),
-      fieldWithPath("activityType").type(JsonFieldType.STRING).description("활동 종류"),
-      fieldWithPath("name").type(JsonFieldType.STRING).description("활동 이름")
+        fieldWithPath("id").type(JsonFieldType.NUMBER).description("활동 식별자"),
+        fieldWithPath("activityType").type(JsonFieldType.STRING).description("활동 종류"),
+        fieldWithPath("name").type(JsonFieldType.STRING).description("활동 이름")
     );
 
     //when & then
-    mockMvc.perform(post("/activities")
-        .contentType(MediaType.APPLICATION_JSON)
-        .content(objectMapper.writeValueAsString(request)))
-      .andExpect(status().isCreated())
-      .andDo(document("add-activity", requestFields, responseFields));
+    mockMvc.perform(post("/admin/activities")
+            .contentType(MediaType.APPLICATION_JSON)
+            .content(objectMapper.writeValueAsString(request)))
+        .andExpect(status().isCreated())
+        .andDo(document("add-activity", requestFields, responseFields));
   }
 }
