+++ backend/emm-sale/src/documentTest/java/com/emmsale/TagApiTest.java
@@ -10,6 +10,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import com.emmsale.admin.tag.api.AdminTagApi;
 import com.emmsale.tag.api.TagApi;
 import com.emmsale.tag.application.dto.TagRequest;
 import com.emmsale.tag.application.dto.TagResponse;
@@ -22,7 +23,7 @@ import org.springframework.restdocs.payload.JsonFieldType;
 import org.springframework.restdocs.payload.RequestFieldsSnippet;
 import org.springframework.restdocs.payload.ResponseFieldsSnippet;
 
-@WebMvcTest(TagApi.class)
+@WebMvcTest({TagApi.class, AdminTagApi.class})
 class TagApiTest extends MockMvcTestHelper {
 
   @Test
@@ -66,7 +67,7 @@ class TagApiTest extends MockMvcTestHelper {
     );
 
     //when & then
-    mockMvc.perform(post("/tags")
+    mockMvc.perform(post("/admin/tags")
             .contentType(MediaType.APPLICATION_JSON)
             .content(objectMapper.writeValueAsString(request)))
         .andExpect(status().isCreated())
