+++ backend/emm-sale/src/main/java/com/emmsale/admin/tag/application/TagCommandService.java
@@ -1,4 +1,4 @@
-package com.emmsale.tag.application;
+package com.emmsale.admin.tag.application;
 
 import com.emmsale.tag.application.dto.TagRequest;
 import com.emmsale.tag.application.dto.TagResponse;
